+++ mountpoint-s3/CHANGELOG.md
@@ -1,11 +1,13 @@
-## Unreleased
+## Unreleased (v1.18.1)
+
+* Update underlying S3 client version.
 
 ## v1.18.0 (May 30, 2025)
 
 ### New features
 
 * Allow launching Mountpoint from the `/etc/fstab` file. See [Automatically mounting an S3 bucket at boot](https://github.com/awslabs/mountpoint-s3/blob/main/doc/CONFIGURATION.md#automatically-mounting-an-s3-bucket-at-boot) for more details ([#1441](https://github.com/awslabs/mountpoint-s3/pull/1441), [#1362](https://github.com/awslabs/mountpoint-s3/pull/1362), [#1434](https://github.com/awslabs/mountpoint-s3/pull/1434))
-* Allow passing in s3:// URIs as the first argument to Mountpoint. The `--prefix` option is not allowed if a prefix is passed in as part of the S3 URI. ([#1434](https://github.com/awslabs/mountpoint-s3/pull/1434))
+* Allow passing in s3:// URIs as the first argument to Mountpoint. The `--prefix` option is not allowed if using an S3 URI. ([#1434](https://github.com/awslabs/mountpoint-s3/pull/1434))
 * Allow passing in s3:// URIs to the `--cache-xz` parameter. Use of prefixes is still unsupported. ([#1434](https://github.com/awslabs/mountpoint-s3/pull/1434))
 
 ## v1.17.0 (May 12, 2025)